requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
llama-index
llama-index-embeddings-huggingface
llama-index-llms-groq
pypdf
python-docx
llama-index-vector-stores-pinecone
pinecone-client
llama-index-postprocessor-cohere-rerank
gunicorn
sentence-transformers
torch